+++ src/components/QuestionForm/QuestionAddForm/index.tsx
@@ -3,7 +3,7 @@ import { useQueryClient } from '@tanstack/react-query';
 import { Button, Input, Textarea } from 'review-me-design-system';
 import { useUserContext } from '@contexts/userContext';
 import { usePostQuestion } from '@apis/questionApi';
-import { ButtonWrapper, KeywordLabel, QuestionFormLayout } from './style';
+import { ButtonWrapper, KeywordLabel, QuestionFormLayout } from '../style';
 
 interface Props {
   resumeId: number;
@@ -50,7 +50,7 @@ const QuestionForm = ({ resumeId, currentPageNum }: Props) => {
   };
 
   return (
-    <QuestionFormLayout onSubmit={handleSubmit}>
+    <QuestionFormLayout $type="add" onSubmit={handleSubmit}>
       <KeywordLabel>{labelContent}</KeywordLabel>
       <Input
         placeholder="예상질문 키워드"
@@ -58,7 +58,7 @@ const QuestionForm = ({ resumeId, currentPageNum }: Props) => {
         onChange={(e) => setLabelContent(e.target.value)}
       />
       <Textarea placeholder="예상질문" value={content} onChange={(e) => setContent(e.target.value)} />
-      <ButtonWrapper>
+      <ButtonWrapper $type="add">
         <Button variant="default" size="s">
           작성
         </Button>
