+++ src/components/QuestionForm/style.ts
@@ -1,14 +1,14 @@
 import { theme } from 'review-me-design-system';
 import styled from 'styled-components';
 
-const QuestionFormLayout = styled.form`
+const QuestionFormLayout = styled.form<{ $type: 'add' | 'edit' }>`
   display: flex;
   flex-direction: column;
   justify-content: center;
   gap: 0.5rem;
-  padding: 0.75rem 1rem;
+  ${({ $type }) => $type === 'add' && 'padding: 0.75rem 1rem;'}
 
-  box-shadow: rgba(0, 0, 0, 0.07) 0 0 1.25rem;
+  ${({ $type }) => $type === 'add' && 'box-shadow: rgba(0, 0, 0, 0.07) 0 0 1.25rem'};
 `;
 
 const LabelList = styled.div`
@@ -18,9 +18,10 @@ const LabelList = styled.div`
   gap: 0.5rem;
 `;
 
-const ButtonWrapper = styled.div`
+const ButtonWrapper = styled.div<{ $type: 'add' | 'edit' }>`
   display: flex;
   justify-content: flex-end;
+  ${({ $type }) => $type === 'edit' && 'gap: 0.5rem;'}
   width: 100%;
 `;
